remove properties no longer common to all categories.
