+++ mes-test-schema/categories/AM1/index.d.ts
@@ -1,0 +1,29 @@
+/* tslint:disable */
+/**
+ * This file was initially generated by json-schema-to-typescript.
+ * IT HAS SINCE BEEN MOFIIED BY HAND. DO NOT REGENERATE
+ */
+
+// Always follow the order of the properties that are in partial.d.ts
+import { PartialTestResultCatAM1Schema } from './partial';
+import { TestResultCommonSchema } from '../common';
+import { TestData as PartialCatAM1TestData } from './partial';
+import { TestData as CommonTestData } from '../common';
+import { JournalData as PartialCatAM1JournalData } from './partial';
+import { JournalData as CommonJournalData } from '../common/index';
+import { Candidate as PartialCatAM1Candidate } from './partial';
+import { Candidate as CommonCandidate } from '../common/index';
+import { VehicleDetails as PartialCatAM1VehicleDetails } from './partial';
+import { VehicleDetails as CommonVehicleDetails } from '../common/index';
+
+export declare namespace CatAM1UniqueTypes {
+    type TestResult = PartialTestResultCatAM1Schema & TestResultCommonSchema;
+
+    type TestData = PartialCatAM1TestData & CommonTestData;
+
+    type JournalData = PartialCatAM1JournalData & CommonJournalData;
+
+    type Candidate = PartialCatAM1Candidate & CommonCandidate;
+
+    type VehicleDetails = PartialCatAM1VehicleDetails & CommonVehicleDetails;
+}
