+++ mes-test-schema/categories/B/index.json
@@ -27,26 +27,26 @@
 				"dangerousFaults": {
 					"$ref": "#/definitions/dangerousFaults"
 				},
-				"eco": {
-					"$ref": "#/definitions/eco"
-				},
-				"ETA": {
-					"$ref": "#/definitions/ETA"
-				},
-				"testRequirements": {
-					"$ref": "#/definitions/testRequirements"
-				},
 				"eyesightTest": {
 					"$ref": "#/definitions/eyesightTest"
 				},
 				"controlledStop": {
 					"$ref": "#/definitions/controlledStop"
 				},
+				"testRequirements": {
+					"$ref": "#/definitions/testRequirements"
+				},
 				"manoeuvres": {
 					"$ref": "#/definitions/manoeuvres"
 				},
 				"vehicleChecks": {
 					"$ref": "#/definitions/vehicleChecks"
+				},
+				"eco": {
+					"$ref": "#/definitions/eco"
+				},
+				"ETA": {
+					"$ref": "#/definitions/ETA"
 				}
 			},
 			"type": "object"
@@ -1292,40 +1292,6 @@
 			},
 			"additionalProperties": false
 		},
-		"eco": {
-			"description": "Assessment of the eco friendly manner of driving",
-			"type": "object",
-			"properties": {
-				"completed": {
-					"description": "Indicates that the eco friendly manner of driving has been assessed",
-					"type": "boolean"
-				},
-				"adviceGivenControl": {
-					"description": "Indicates that advice was given on the Control aspect of eco driving",
-					"type": "boolean"
-				},
-				"adviceGivenPlanning": {
-					"description": "Indicates that advice was given on the Planning aspect of eco driving",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
-		"ETA": {
-			"description": "Indicates whether the examiner had to take physical or verbal action during the test",
-			"type": "object",
-			"properties": {
-				"physical": {
-					"description": "Indicates that the examiner had to take physical action",
-					"type": "boolean"
-				},
-				"verbal": {
-					"description": "Indicates that the examiner had to take verbal action",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
 		"postTestDeclarations": {
 			"type": "object",
 			"properties": {
@@ -1584,29 +1550,6 @@
 			},
 			"additionalProperties": false
 		},
-		"testRequirements": {
-			"description": "The test requirements that must be carried out during a test",
-			"type": "object",
-			"properties": {
-				"normalStart1": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"normalStart2": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"angledStart": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"hillStart": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
 		"vehicleDetails": {
 			"description": "Details about the vehicle to be used for the test",
 			"type": "object",
@@ -1648,6 +1591,40 @@
 				"provisionalLicenceProvided",
 				"passCertificateNumber"
 			]
+		},
+		"eco": {
+			"description": "Assessment of the eco friendly manner of driving",
+			"type": "object",
+			"properties": {
+				"completed": {
+					"description": "Indicates that the eco friendly manner of driving has been assessed",
+					"type": "boolean"
+				},
+				"adviceGivenControl": {
+					"description": "Indicates that advice was given on the Control aspect of eco driving",
+					"type": "boolean"
+				},
+				"adviceGivenPlanning": {
+					"description": "Indicates that advice was given on the Planning aspect of eco driving",
+					"type": "boolean"
+				}
+			},
+			"additionalProperties": false
+		},
+		"ETA": {
+			"description": "Indicates whether the examiner had to take physical or verbal action during the test",
+			"type": "object",
+			"properties": {
+				"physical": {
+					"description": "Indicates that the examiner had to take physical action",
+					"type": "boolean"
+				},
+				"verbal": {
+					"description": "Indicates that the examiner had to take verbal action",
+					"type": "boolean"
+				}
+			},
+			"additionalProperties": false
 		},
 		"controlledStop": {
 			"additionalProperties": false,
@@ -1693,6 +1670,29 @@
 				},
 				"reverseRight": {
 					"$ref": "#/definitions/manoeuvre"
+				}
+			}
+		},
+		"testRequirements": {
+			"description": "The test requirements that must be carried out during a test",
+			"additionalProperties": false,
+			"type": "object",
+			"properties": {
+				"normalStart1": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"normalStart2": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"angledStart": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"hillStart": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
 				}
 			}
 		},
