+++ mes-test-schema/categories/B/partial.d.ts
@@ -68,6 +68,8 @@ export interface TestData {
   testRequirements?: TestRequirements;
   manoeuvres?: Manoeuvres;
   vehicleChecks?: VehicleChecks;
+  eco?: Eco;
+  ETA?: ETA;
 }
 /**
  * This interface was referenced by `PartialTestResultCatBSchema`'s JSON-Schema
@@ -91,10 +93,20 @@ export interface ControlledStop {
   selected?: ManoeuvreIndicator;
 }
 /**
+ * The test requirements that must be carried out during a test
+ *
  * This interface was referenced by `PartialTestResultCatBSchema`'s JSON-Schema
  * via the `definition` "testRequirements".
  */
 export interface TestRequirements {
+  /**
+   * Indicates whether or not this test requirement was carried out
+   */
+  normalStart1?: boolean;
+  /**
+   * Indicates whether or not this test requirement was carried out
+   */
+  normalStart2?: boolean;
   /**
    * Indicates whether or not this test requirement was carried out
    */
@@ -159,6 +171,42 @@ export interface QuestionResult {
    * Outcome of the question that was asked
    */
   outcome?: "P" | "DF" | "S" | "D";
+}
+/**
+ * Assessment of the eco friendly manner of driving
+ *
+ * This interface was referenced by `PartialTestResultCatBSchema`'s JSON-Schema
+ * via the `definition` "eco".
+ */
+export interface Eco {
+  /**
+   * Indicates that the eco friendly manner of driving has been assessed
+   */
+  completed?: boolean;
+  /**
+   * Indicates that advice was given on the Control aspect of eco driving
+   */
+  adviceGivenControl?: boolean;
+  /**
+   * Indicates that advice was given on the Planning aspect of eco driving
+   */
+  adviceGivenPlanning?: boolean;
+}
+/**
+ * Indicates whether the examiner had to take physical or verbal action during the test
+ *
+ * This interface was referenced by `PartialTestResultCatBSchema`'s JSON-Schema
+ * via the `definition` "ETA".
+ */
+export interface ETA {
+  /**
+   * Indicates that the examiner had to take physical action
+   */
+  physical?: boolean;
+  /**
+   * Indicates that the examiner had to take verbal action
+   */
+  verbal?: boolean;
 }
 /**
  * This interface was referenced by `PartialTestResultCatBSchema`'s JSON-Schema
