+++ mes-test-schema/categories/C/index.json
@@ -27,12 +27,6 @@
 				"dangerousFaults": {
 					"$ref": "#/definitions/dangerousFaults"
 				},
-				"eco": {
-					"$ref": "#/definitions/eco"
-				},
-				"ETA": {
-					"$ref": "#/definitions/ETA"
-				},
 				"testRequirements": {
 					"$ref": "#/definitions/testRequirements"
 				},
@@ -41,6 +35,12 @@
 				},
 				"vehicleChecks": {
 					"$ref": "#/definitions/vehicleChecks"
+				},
+				"eco": {
+					"$ref": "#/definitions/eco"
+				},
+				"ETA": {
+					"$ref": "#/definitions/ETA"
 				}
 			},
 			"type": "object"
@@ -1296,40 +1296,6 @@
 			},
 			"additionalProperties": false
 		},
-		"eco": {
-			"description": "Assessment of the eco friendly manner of driving",
-			"type": "object",
-			"properties": {
-				"completed": {
-					"description": "Indicates that the eco friendly manner of driving has been assessed",
-					"type": "boolean"
-				},
-				"adviceGivenControl": {
-					"description": "Indicates that advice was given on the Control aspect of eco driving",
-					"type": "boolean"
-				},
-				"adviceGivenPlanning": {
-					"description": "Indicates that advice was given on the Planning aspect of eco driving",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
-		"ETA": {
-			"description": "Indicates whether the examiner had to take physical or verbal action during the test",
-			"type": "object",
-			"properties": {
-				"physical": {
-					"description": "Indicates that the examiner had to take physical action",
-					"type": "boolean"
-				},
-				"verbal": {
-					"description": "Indicates that the examiner had to take verbal action",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
 		"postTestDeclarations": {
 			"type": "object",
 			"properties": {
@@ -1588,33 +1554,6 @@
 			},
 			"additionalProperties": false
 		},
-		"testRequirements": {
-			"description": "The test requirements that must be carried out during a test",
-			"type": "object",
-			"properties": {
-				"normalStart1": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"normalStart2": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"angledStartControlledStop": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"downhillStart": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"uphillStart": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
 		"vehicleDetails": {
 			"description": "Details about the vehicle to be used for the test",
 			"type": "object",
@@ -1662,6 +1601,40 @@
 				"passCertificateNumber"
 			]
 		},
+		"eco": {
+			"description": "Assessment of the eco friendly manner of driving",
+			"type": "object",
+			"properties": {
+				"completed": {
+					"description": "Indicates that the eco friendly manner of driving has been assessed",
+					"type": "boolean"
+				},
+				"adviceGivenControl": {
+					"description": "Indicates that advice was given on the Control aspect of eco driving",
+					"type": "boolean"
+				},
+				"adviceGivenPlanning": {
+					"description": "Indicates that advice was given on the Planning aspect of eco driving",
+					"type": "boolean"
+				}
+			},
+			"additionalProperties": false
+		},
+		"ETA": {
+			"description": "Indicates whether the examiner had to take physical or verbal action during the test",
+			"type": "object",
+			"properties": {
+				"physical": {
+					"description": "Indicates that the examiner had to take physical action",
+					"type": "boolean"
+				},
+				"verbal": {
+					"description": "Indicates that the examiner had to take verbal action",
+					"type": "boolean"
+				}
+			},
+			"additionalProperties": false
+		},
 		"manoeuvres": {
 			"additionalProperties": false,
 			"description": "The manoeuvres that were carried out during the test and any faults recorded against them",
@@ -1703,6 +1676,33 @@
 			"description": "Telephone number of the business the candidate relates to",
 			"type": "string",
 			"maxLength": 20
+		},
+		"testRequirements": {
+			"additionalProperties": false,
+			"description": "The test requirements that must be carried out during a test",
+			"properties": {
+				"normalStart1": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"normalStart2": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"angledStartControlledStop": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"downhillStart": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"uphillStart": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				}
+			},
+			"type": "object"
 		}
 	},
 	"properties": {
