+++ mes-test-schema/categories/C1E/index.json
@@ -27,23 +27,23 @@
 				"dangerousFaults": {
 					"$ref": "#/definitions/dangerousFaults"
 				},
-				"eco": {
-					"$ref": "#/definitions/eco"
-				},
-				"ETA": {
-					"$ref": "#/definitions/ETA"
+				"uncoupleRecouple": {
+					"$ref": "#/definitions/uncoupleRecouple"
 				},
 				"testRequirements": {
 					"$ref": "#/definitions/testRequirements"
-				},
-				"uncoupleRecouple": {
-					"$ref": "#/definitions/uncoupleRecouple"
 				},
 				"manoeuvres": {
 					"$ref": "#/definitions/manoeuvres"
 				},
 				"vehicleChecks": {
 					"$ref": "#/definitions/vehicleChecks"
+				},
+				"eco": {
+					"$ref": "#/definitions/eco"
+				},
+				"ETA": {
+					"$ref": "#/definitions/ETA"
 				}
 			},
 			"type": "object"
@@ -1299,40 +1299,6 @@
 			},
 			"additionalProperties": false
 		},
-		"eco": {
-			"description": "Assessment of the eco friendly manner of driving",
-			"type": "object",
-			"properties": {
-				"completed": {
-					"description": "Indicates that the eco friendly manner of driving has been assessed",
-					"type": "boolean"
-				},
-				"adviceGivenControl": {
-					"description": "Indicates that advice was given on the Control aspect of eco driving",
-					"type": "boolean"
-				},
-				"adviceGivenPlanning": {
-					"description": "Indicates that advice was given on the Planning aspect of eco driving",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
-		"ETA": {
-			"description": "Indicates whether the examiner had to take physical or verbal action during the test",
-			"type": "object",
-			"properties": {
-				"physical": {
-					"description": "Indicates that the examiner had to take physical action",
-					"type": "boolean"
-				},
-				"verbal": {
-					"description": "Indicates that the examiner had to take verbal action",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
 		"postTestDeclarations": {
 			"type": "object",
 			"properties": {
@@ -1591,33 +1557,6 @@
 			},
 			"additionalProperties": false
 		},
-		"testRequirements": {
-			"description": "The test requirements that must be carried out during a test",
-			"type": "object",
-			"properties": {
-				"normalStart1": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"normalStart2": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"angledStartControlledStop": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"downhillStart": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				},
-				"uphillStart": {
-					"description": "Indicates whether or not this test requirement was carried out",
-					"type": "boolean"
-				}
-			},
-			"additionalProperties": false
-		},
 		"vehicleDetails": {
 			"description": "Details about the vehicle to be used for the test",
 			"type": "object",
@@ -1661,6 +1600,40 @@
 				"provisionalLicenceProvided",
 				"passCertificateNumber"
 			]
+		},
+		"eco": {
+			"description": "Assessment of the eco friendly manner of driving",
+			"type": "object",
+			"properties": {
+				"completed": {
+					"description": "Indicates that the eco friendly manner of driving has been assessed",
+					"type": "boolean"
+				},
+				"adviceGivenControl": {
+					"description": "Indicates that advice was given on the Control aspect of eco driving",
+					"type": "boolean"
+				},
+				"adviceGivenPlanning": {
+					"description": "Indicates that advice was given on the Planning aspect of eco driving",
+					"type": "boolean"
+				}
+			},
+			"additionalProperties": false
+		},
+		"ETA": {
+			"description": "Indicates whether the examiner had to take physical or verbal action during the test",
+			"type": "object",
+			"properties": {
+				"physical": {
+					"description": "Indicates that the examiner had to take physical action",
+					"type": "boolean"
+				},
+				"verbal": {
+					"description": "Indicates that the examiner had to take verbal action",
+					"type": "boolean"
+				}
+			},
+			"additionalProperties": false
 		},
 		"uncoupleRecouple": {
 			"additionalProperties": false,
@@ -1718,6 +1691,33 @@
 			"description": "Telephone number of the business the candidate relates to",
 			"type": "string",
 			"maxLength": 20
+		},
+		"testRequirements": {
+			"additionalProperties": false,
+			"description": "The test requirements that must be carried out during a test",
+			"properties": {
+				"normalStart1": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"normalStart2": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"angledStartControlledStop": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"downhillStart": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				},
+				"uphillStart": {
+					"description": "Indicates whether or not this test requirement was carried out",
+					"type": "boolean"
+				}
+			},
+			"type": "object"
 		}
 	},
 	"properties": {
