+++ mes-test-schema/category-definitions/C1/partial.json
@@ -2,6 +2,40 @@
   "title": "Partial Test Result Cat C1 Schema",
   "type": "object",
   "definitions": {
+    "eco": {
+      "description": "Assessment of the eco friendly manner of driving",
+      "type": "object",
+      "properties": {
+        "completed": {
+          "description": "Indicates that the eco friendly manner of driving has been assessed",
+          "type": "boolean"
+        },
+        "adviceGivenControl": {
+          "description": "Indicates that advice was given on the Control aspect of eco driving",
+          "type": "boolean"
+        },
+        "adviceGivenPlanning": {
+          "description": "Indicates that advice was given on the Planning aspect of eco driving",
+          "type": "boolean"
+        }
+      },
+      "additionalProperties": false
+    },
+    "ETA": {
+      "description": "Indicates whether the examiner had to take physical or verbal action during the test",
+      "type": "object",
+      "properties": {
+        "physical": {
+          "description": "Indicates that the examiner had to take physical action",
+          "type": "boolean"
+        },
+        "verbal": {
+          "description": "Indicates that the examiner had to take verbal action",
+          "type": "boolean"
+        }
+      },
+      "additionalProperties": false
+    },
     "faultComments": {
       "$ref": "../common/index.json#/definitions/faultComments"
     },
@@ -114,6 +148,12 @@
         },
         "vehicleChecks": {
           "$ref": "#/definitions/vehicleChecks"
+        },
+        "eco": {
+          "$ref": "#/definitions/eco"
+        },
+        "ETA": {
+          "$ref": "#/definitions/ETA"
         }
       },
       "additionalProperties": false
@@ -122,6 +162,14 @@
       "additionalProperties": false,
       "description": "The test requirements that must be carried out during a test",
       "properties": {
+        "normalStart1": {
+          "description": "Indicates whether or not this test requirement was carried out",
+          "type": "boolean"
+        },
+        "normalStart2": {
+          "description": "Indicates whether or not this test requirement was carried out",
+          "type": "boolean"
+        },
         "angledStartControlledStop": {
           "description": "Indicates whether or not this test requirement was carried out",
           "type": "boolean"
