+++ mes-test-schema/category-definitions/CE/partial.json
@@ -2,6 +2,40 @@
   "title": "Partial Test Result Cat CE Schema",
   "type": "object",
   "definitions": {
+    "eco": {
+      "description": "Assessment of the eco friendly manner of driving",
+      "type": "object",
+      "properties": {
+        "completed": {
+          "description": "Indicates that the eco friendly manner of driving has been assessed",
+          "type": "boolean"
+        },
+        "adviceGivenControl": {
+          "description": "Indicates that advice was given on the Control aspect of eco driving",
+          "type": "boolean"
+        },
+        "adviceGivenPlanning": {
+          "description": "Indicates that advice was given on the Planning aspect of eco driving",
+          "type": "boolean"
+        }
+      },
+      "additionalProperties": false
+    },
+    "ETA": {
+      "description": "Indicates whether the examiner had to take physical or verbal action during the test",
+      "type": "object",
+      "properties": {
+        "physical": {
+          "description": "Indicates that the examiner had to take physical action",
+          "type": "boolean"
+        },
+        "verbal": {
+          "description": "Indicates that the examiner had to take verbal action",
+          "type": "boolean"
+        }
+      },
+      "additionalProperties": false
+    },
     "passCompletion": {
       "additionalProperties": false,
       "properties": {
@@ -143,6 +177,12 @@
         },
         "vehicleChecks": {
           "$ref": "#/definitions/vehicleChecks"
+        },
+        "eco": {
+          "$ref": "#/definitions/eco"
+        },
+        "ETA": {
+          "$ref": "#/definitions/ETA"
         }
       },
       "additionalProperties": false
@@ -151,6 +191,14 @@
       "additionalProperties": false,
       "description": "The test requirements that must be carried out during a test",
       "properties": {
+        "normalStart1": {
+          "description": "Indicates whether or not this test requirement was carried out",
+          "type": "boolean"
+        },
+        "normalStart2": {
+          "description": "Indicates whether or not this test requirement was carried out",
+          "type": "boolean"
+        },
         "angledStartControlledStop": {
           "description": "Indicates whether or not this test requirement was carried out",
           "type": "boolean"
