+++ mes-test-schema/category-definitions/common/index.json
@@ -26,15 +26,6 @@
         },
         "dangerousFaults": {
           "$ref": "#/definitions/dangerousFaults"
-        },
-        "eco": {
-          "$ref": "#/definitions/eco"
-        },
-        "ETA": {
-          "$ref": "#/definitions/ETA"
-        },
-        "testRequirements": {
-          "$ref": "#/definitions/testRequirements"
         }
       },
       "type": "object"
@@ -1280,40 +1271,6 @@
       },
       "additionalProperties": false
     },
-    "eco": {
-      "description": "Assessment of the eco friendly manner of driving",
-      "type": "object",
-      "properties": {
-        "completed": {
-          "description": "Indicates that the eco friendly manner of driving has been assessed",
-          "type": "boolean"
-        },
-        "adviceGivenControl": {
-          "description": "Indicates that advice was given on the Control aspect of eco driving",
-          "type": "boolean"
-        },
-        "adviceGivenPlanning": {
-          "description": "Indicates that advice was given on the Planning aspect of eco driving",
-          "type": "boolean"
-        }
-      },
-      "additionalProperties": false
-    },
-    "ETA": {
-      "description": "Indicates whether the examiner had to take physical or verbal action during the test",
-      "type": "object",
-      "properties": {
-        "physical": {
-          "description": "Indicates that the examiner had to take physical action",
-          "type": "boolean"
-        },
-        "verbal": {
-          "description": "Indicates that the examiner had to take verbal action",
-          "type": "boolean"
-        }
-      },
-      "additionalProperties": false
-    },
     "postTestDeclarations": {
       "type": "object",
       "properties": {
@@ -1568,21 +1525,6 @@
         "reason": {
           "description": "The reason this option was selected",
           "type": "string"
-        }
-      },
-      "additionalProperties": false
-    },
-    "testRequirements": {
-      "description": "The test requirements that must be carried out during a test",
-      "type": "object",
-      "properties": {
-        "normalStart1": {
-          "description": "Indicates whether or not this test requirement was carried out",
-          "type": "boolean"
-        },
-        "normalStart2": {
-          "description": "Indicates whether or not this test requirement was carried out",
-          "type": "boolean"
         }
       },
       "additionalProperties": false
